Make port settings for Charon consistent for internal/external
If the ports which the Charon service runs on (and are connected to on the container side) do not match the ports that are bound to the host, then the incorrect port will be advertised to 'some' Charon peers, which (so we found) can lead to peering issues in 'some' cases. (EG nodes which are colocated with one another)

This may hinder any efforts to add redundancy within a single site, making any 'local' redundancy efforts a semi-pointless consideration compared to only having distributed redundancy. (EG, 2+ nodes per site, is better than 1 node per site, right?)